cleanup  components folder
