+++ src/components/svg/svg.container.js
