--- src/css/layout.css
@@ -1,23 +1,0 @@
-/*.flex {
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  flex-wrap: wrap;
-}*/
-.container {
-  margin: 2rem; }
-
-.grow {
-  flex-grow: 1; }
-
-.baseline {
-  align-items: baseline; }
-
-.relative {
-  position: relative; }
-
-.absolute {
-  position: absolute; }
-  .absolute.right {
-    top: 0;
-    right: 0; }
