--- src/css/profile.container.css
@@ -1,32 +1,0 @@
-.profile {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  margin: 1rem;
-  padding: 2rem;
-  width: 300px;
-  flex-grow: 0;
-  flex-shrink: 0;
-  background-color: white;
-  box-shadow: 1px 1px 5px rgba(170, 170, 170, 0.35);
-  border-radius: 2px; }
-  .profile .profile-buttons {
-    display: flex;
-    justify-content: space-between; }
-
-input {
-  width: 100%;
-  display: inline-block;
-  border: none;
-  outline: 0;
-  color: white;
-  transition: all 200ms;
-  font-size: medium;
-  background-color: transparent; }
-
-input.underline {
-  color: #292b2c;
-  border-bottom: 2px solid #292b2c; }
-
-input.underline:focus {
-  border-bottom-color: #296bbd; }
