--- src/css/recommender.css
@@ -1,7 +1,0 @@
-.recommender-buttons {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap; }
-  .recommender-buttons a {
-    flex-grow: 1;
-    margin: 1rem; }
